assistant/requirements.txt
accelerate==0.25.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
boto3==1.34.49
botocore==1.34.49
cached_path==1.6.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
dateparser==1.1.8
decorator==5.1.1
diskcache==5.6.3
docopt==0.6.2
einops==0.7.0
einops-exts==0.0.4
filelock==3.12.4
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gruut==2.3.4
gruut-ipa==0.13.0
gruut_lang_en==2.0.0
halo==0.0.31
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-experimental==0.0.52
langsmith==0.1.7
lazy_loader==0.3
librosa==0.10.1
llama_cpp_python==0.2.50
llvmlite==0.42.0
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
munch==4.0.0
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
orjson==3.9.15
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
psutil==5.9.8
pvporcupine==3.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.1
spinners==0.0.24
SQLAlchemy==2.0.27
styletts2==0.1.6
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
torchaudio==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.0.0
yarl==1.9.4
